Consolidate utils/util.
