+++ api/util.py
@@ -2,12 +2,18 @@
 Utility/helper functions.
 """
 
+import re
+import aiodns
 import datetime
 import hashlib
 import random
 import string
 import time
 import orjson as json
+from typing import Set
+from ipaddress import ip_address, IPv4Address, IPv6Address
+
+ALLOWED_HOST_RE = re.compile(r"(?!-)[a-z\d-]{1,63}(?<!-)$")
 
 
 def now_str():
@@ -54,3 +60,68 @@ def get_signing_message(
         return f"{hotkey}:{nonce}:{payload_hash}"
     else:
         raise ValueError("Either payload_str or purpose must be provided")
+
+
+def is_invalid_ip(ip: IPv4Address | IPv6Address) -> bool:
+    """
+    Check if IP address is private/local network.
+    """
+    return (
+        ip.is_private
+        or ip.is_loopback
+        or ip.is_link_local
+        or ip.is_multicast
+        or ip.is_reserved
+        or ip.is_unspecified
+    )
+
+
+async def get_resolved_ips(host: str) -> Set[IPv4Address | IPv6Address]:
+    """
+    Resolve all IP addresses for a host.
+    """
+    resolver = aiodns.DNSResolver()
+    resolved_ips = set()
+    try:
+        # IPv4
+        try:
+            result = await resolver.query(host, "A")
+            for answer in result:
+                resolved_ips.add(ip_address(answer.host))
+        except aiodns.error.DNSError:
+            pass
+
+        # IPv6
+        try:
+            result = await resolver.query(host, "AAAA")
+            for answer in result:
+                resolved_ips.add(ip_address(answer.host))
+        except aiodns.error.DNSError:
+            pass
+        if not resolved_ips:
+            raise ValueError(f"Could not resolve any IP addresses for host: {host}")
+        return resolved_ips
+    except Exception as exc:
+        raise ValueError(f"DNS resolution failed for host {host}: {str(exc)}")
+
+
+async def is_valid_host(host: str) -> bool:
+    """
+    Validate host (IP or DNS name).
+    """
+    if not host or len(host) > 255:
+        return False
+    if not all(ALLOWED_HOST_RE.match(x) for x in host.lower().rstrip(".").split(".")):
+        return False
+    try:
+        # IP address provided.
+        addr = ip_address(host)
+        return not is_invalid_ip(addr)
+    except ValueError:
+        # DNS hostname provided, look up IPs.
+        try:
+            resolved_ips = await get_resolved_ips(host)
+            return all(not is_invalid_ip(ip) for ip in resolved_ips)
+        except ValueError:
+            return False
+    return False
